Create CODE_OF_CONDUCT.md
Is this the right email address to send complaints?  Would it feel better and be more appropriate to create OpenEBS email addresses?
More importantly - this is the default and useful looking code of conduct recommended by GitHub.  As such, unless we feel strongly otherwise I think we should use it.