Setup kube-proxy in user-mode networking Fixes #302
As part of the recent upgrades, the changes (PR #295) to set the
kube-proxy in user mode were reverted. As it turns out, the issue
of weave pods getting stuck in CrashLoopBackOff appears intermittently
and on restarts of the kubeminion nodes.

Reverting back the changes that removed setting kube-proxy in
user-mode