Framework for patching existing vagrant boxes
There are scenarios where, once an box is uploaded,
the box needs to be updated with either files or
packages.

In this case, the configure_k8s_weave.sh was required
to be updated.

This commit provides an initial framework using which
the box can be upgraded. This commit hard-codes the
version of the box to be patched and the patch code
itself. There will be enhancment requests raised to
make this more generic.