Update CAS Templates to configure Resource Requests (#1851)
* Update CAS Templates to configure Resource Requests

When ResourceLimits are specified, the same values are used
as requests by Kubernetes. For example, when a limit of 4Gi
is set on a container, the limit is also set as 4Gi. This can
cause the pods to be unschedulable if nodes can't satisfy
during scheduling. While this may be acceptable for auxillary
(or side car) containers, for cpu/memory intensive pods like
- pool, target or replica needs to be schedulable by setting
a upper limit that could usually be more than other typical
pods.

This PR adds the support to pass in the Resource Requests to
CAS templates.

Also, the default cstor sparse pool values have been updated
to limit the memory consumption to 2Gi max.

Signed-off-by: kmova <kiran.mova@openebs.io>

* Update the comment for default sparse SPC

Signed-off-by: kmova <kiran.mova@openebs.io>