docs(contributor): Propose format for openebs git commit messages
OpenEBS Code is split across multiple repositories. The current
commit messages are not uniform across and within the repo.
This makes it hard to generate release notes or to onboard
new contributors.

This document proposes a convention that can be followed
for git messages across different OpenEBS repositories.
Following the convention allows for automating
Release Note generation as well as improving the readability.

Signed-off-by: kmova <kiran.mova@openebs.io>