refactor(fio): split the SC and SPC from Pod YAML
Some of the sample fio pod YAMLs contain the
corresponding SC or SPC within them. If the
user hits - `kubectl delete -f ` of Pod YAML
and if the associated SPC or SC are deleted,
the PV related pods associated with the Pod
will become orphaned.

Keeping the SPC and SC in different YAML
files than Pod YAML.

Signed-off-by: kmova <kiran.mova@openebs.io>