feat(perf): sample dbench files for benchmark tests
This PR includes some sample YAMLs using the
logdna/dbench fio scripts for running bench marking
tests using cstor and jiva engines. Also includes
YAMLs for running with host-path volumes.

This is an initial commit of the files that can be
further refined as new tunables are added in the
coming versions

Signed-off-by: kmova <kiran.mova@openebs.io>