docs(contributor): add governance document
The top contributors for the OpenEBS projects are
from MayaData and Viking, and the current governance was primarily
done via a syncup on the community slack or with direct interactions.
As the project is gaining momentum and interest from more
companies to contribute, it is time to outline the
current governance model.

After looking at the various examples provided in https://github.com/openebs/openebs/issues/2545,
the closest that matches OpenEBS Governance mode is from:
https://github.com/containerd/project/blob/master/GOVERNANCE.md

The text in this document is borrowed from the above containerd
GOVERNANCE doc.

Please feel free to suggest changes and corrections.

Signed-off-by: kmova <kiran.mova@openebs.io>