feat(upgrade, ndm): add patch files for disk upgrade (#2641)
The disk CR has been modified to include filesystem and mountpoint information. In the process, the spec.filesystem changes from being a string to struct. This causes the latest NDM pods to remain in stuck state, when reading existing disk CRs from older versions. 

This PR removes the older struct which is unused. After the restart, the NDM pods will update with the latest struct.

Signed-off-by: mittachaitu <sai.chaithanya@mayadata.io>