fix(examples): update the benchmarking docker image
The image (logdna/dbench) is no longer accessible. The
same image has been built and pushed from the
https://github.com/openebs/fbench ( as openebs/fbench)

This PR updates the YAMLs to use openebs/fbench.
The image is renamed from dbench to fbench as it is
based on the fio tool.

This PR also includes an example to run the benchmark
tests on local pv with hostdevice type.

Signed-off-by: kmova <kiran.mova@mayadata.io>