feat(governance): add Praveen as new reviewer
Refer: [GOVERNANCE.md](https://github.com/openebs/openebs/blob/master/GOVERNANCE.md).

Praveen has been helping with developing LVM and ZFS LocalPV
storage engines of the OpenEBS project that broadly fall
in the purview of the control plane.

Based on his interest/commitment to contributing to OpenEBS
and active involvement in the OpenEBS CSI community meetings,
we take pride in adding him as a reviewer to the OpenEBS project.

Signed-off-by: Pawan <pawan@mayadata.io>